Split known response item metadata
